Solved all linting issues + disabled import/no-resolve eslint-rule
I refactored the codebase to comply with all the ESlint-rules. This
mainly consisted of deleting unused imports, destrucuring variables and
implementing type checking for props. I did have to disable the
import/no-resolve eslint rule because it constanntly gave me false
positives when trying to use es6-import syntax. BUG: when I try to
change the name of the jsx-element (variable name) on the contact page
the page starts to crash. How do change the variable name of the page
without this happening?